add split brain messages
